refactor: move UserMapper to another file feat: implement FriendListDao
