+++ src/main/java/ru/yandex/practicum/filmorate/dao/FriendListDao.java
@@ -1,5 +1,7 @@
 package ru.yandex.practicum.filmorate.dao;
 
+import ru.yandex.practicum.filmorate.model.User;
+
 import java.util.List;
 
 public interface FriendListDao {
@@ -8,5 +10,5 @@ public interface FriendListDao {
 
     void removeFriend(Long userId, Long friendId);
 
-    List<Long> getFriends(Long userId);
+    List<User> getFriends(Long userId);
 }
