+++ src/main/java/ru/yandex/practicum/filmorate/dao/impl/FriendListDb.java
@@ -1,31 +1,38 @@
 package ru.yandex.practicum.filmorate.dao.impl;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.dao.FriendListDao;
+import ru.yandex.practicum.filmorate.dao.mapper.UserMapper;
+import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.List;
 
 @Component
+@RequiredArgsConstructor
 public class FriendListDb implements FriendListDao {
 
-    JdbcTemplate jdbcTemplate;
+    private final JdbcTemplate jdbcTemplate;
 
     @Override
     public void addFriend(Long userId, Long friendId) {
-        String query = "INSERT INTO friends_list (user_id, friend_id, state) VALUES (?, ?, FALSE)";
+        String query = "INSERT INTO friend_list (user_id, friend_id, state) VALUES (?, ?, FALSE)";
         jdbcTemplate.update(query, userId, friendId);
     }
 
     @Override
     public void removeFriend(Long userId, Long friendId) {
-        String query = "DELETE FROM friends_list WHERE user_id = ? AND friends_id = ?";
+        String query = "DELETE FROM friend_list WHERE user_id = ? AND friend_id = ?";
         jdbcTemplate.update(query, userId, friendId);
     }
 
     @Override
-    public List<Long> getFriends(Long userId) {
-        String query = "SELECT friends_id FROM friends_list WHERE user_id = ?";
-        return jdbcTemplate.query(query, (rs, rowNum) -> rs.getLong("friend_id"), userId);
+    public List<User> getFriends(Long userId) {
+        String query = "SELECT * FROM users " +
+                "WHERE id IN(SELECT fl.friend_id " +
+                             "FROM friend_list AS fl " +
+                             "WHERE fl.user_id = ?)";
+        return jdbcTemplate.query(query, new UserMapper(), userId);
     }
 }
