+++ src/main/java/ru/yandex/practicum/filmorate/dao/impl/UserDbStorage.java
@@ -5,14 +5,12 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Primary;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.dao.UserDao;
+import ru.yandex.practicum.filmorate.dao.mapper.UserMapper;
 import ru.yandex.practicum.filmorate.model.User;
 
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -79,16 +77,4 @@ public class UserDbStorage implements UserDao {
         return parameters;
     }
 
-    private static final class UserMapper implements RowMapper<User>{
-
-        @Override
-        public User mapRow(ResultSet rs, int rowNum) throws SQLException {
-            return User.builder()
-                    .id(rs.getLong("id"))
-                    .email(rs.getString("email"))
-                    .login(rs.getString("login"))
-                    .name(rs.getString("name"))
-                    .birthday(rs.getDate("birthday").toLocalDate()).build();
-        }
-    }
 }
