+++ src/main/java/ru/yandex/practicum/filmorate/dao/mapper/UserMapper.java
@@ -1,0 +1,21 @@
+package ru.yandex.practicum.filmorate.dao.mapper;
+
+
+import org.springframework.jdbc.core.RowMapper;
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.sql.ResultSet;
+import java.sql.SQLException;
+
+public final class UserMapper implements RowMapper<User> {
+
+    @Override
+    public User mapRow(ResultSet rs, int rowNum) throws SQLException {
+        return User.builder()
+                .id(rs.getLong("id"))
+                .email(rs.getString("email"))
+                .login(rs.getString("login"))
+                .name(rs.getString("name"))
+                .birthday(rs.getDate("birthday").toLocalDate()).build();
+    }
+}
